Distribute files from util/
- fileUtils.js was mostly moved to shared/open.js and the module loading
  part went into appLoader.js
- appLoader.js was mostly moved to app/ and some parts were extracted into
  shared/appDirs.js
- systemReport.js was moved to shared/
