+++ src/app/index.jsx
@@ -42,9 +42,9 @@ import './module-loader';
 import React from 'react';
 import { remote } from 'electron';
 import ReactDOM from 'react-dom';
-import { initAppDirectories, getAppLogDir, loadApp } from '../util/appLoader';
+import { initAppDirectories, loadApp } from './appLoader';
 import legacyRenderer from '../legacy/legacyRenderer';
-import { logger } from '../shared';
+import { logger, getAppLogDir } from '../shared';
 
 const params = new URL(window.location).searchParams;
 const appPath = params.get('appPath');
